Fix Sourcetrail compilation for LLVM-15 and Boost 1.81.
Added a hack in CMakeLists.txt to have a successful link on Ubuntu.
This should be revisited. Probably the compile/link options are
not 100% right for the linker to find all libraries.

See also 2023-04-01-Ubuntu-build.pdf for my Ubuntu build instructions
(including source changes committed here, and this link hack).